services/vanna/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
groq==0.4.1
psycopg[binary]==3.1.18
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
gunicorn==21.2.0
